+++ frontend/src/views/Reporting/ReportingComponents/reportSections/index.ts
@@ -1,0 +1,7 @@
+export { default as generateApplicationSchedule } from './applicationSchedule';
+export { default as generateManureInventory } from './manureInventory';
+export { default as generateManureAndCompostUse } from './manureAndCompostUse';
+export { default as generateFertilizerRequired } from './fertilizerRequired';
+export { default as generateFieldSummary } from './fieldSummary';
+export { default as generateManureAndCompostAnalysis } from './manureAndCompostAnalysis';
+export { default as generateSoilTestResults } from './soilTestResults';
